% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pzfx_tables.R
\name{pzfx_tables}
\alias{pzfx_tables}
\title{List all tables in a GraphPad Prism .pzfx file.}
\usage{
pzfx_tables(path)
}
\arguments{
\item{path}{Path to the .pzfx file}
}
\value{
a character string vector
}
\description{
List all tables in a GraphPad Prism .pzfx file.
}
\examples{
\dontrun{
pzfx_tables(path)
}

}
