% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pzfx.R
\name{read_pzfx}
\alias{read_pzfx}
\title{Read pzfx files}
\usage{
read_pzfx(path, table = 1, strike_action = "exclude", tidify = FALSE)
}
\arguments{
\item{path}{Path to the pzfx file.}

\item{table}{Table to read. Either a string (the name of a table), or an
integer (the position of the table). If neither argument specifies the
table, defaults to the first table.}

\item{strike_action}{One of c("exclude", "keep", "star") or c("e", "k", "s"). Should striked
values in the original pzfx be excluded, kept or labeled with a trailing "*". If a trailing
"*" is added, the column will be of type character.}

\item{tidify}{Logical. Should output data frame be tidified. Currently
not implemented.}
}
\value{
a data frame
}
\description{
Read pzfx files
}
\examples{
\dontrun{
read_pzfx(path, table = 1)
}

}
